Create Probability of Detection plot for GHGSat
Required me to add a new column to clean data indicating if the operator kept the overpass or not ("OperatorKeep"). I now use this as a filter in the data manipulation methods (apply_qc_filter, the method that merges operator report and operator meter data and applies Stanford's QC filter.